Some cleanup.
git-svn-id: https://desktop-5lsudsp:8443/svn/RTCW-1.0@13 a6496cbe-712b-c041-b8d2-258aa9342960

+++ trunk/src/jpeg-6/jmorecfg.h
@@ -149,7 +149,10 @@ typedef unsigned short UINT16;
 typedef unsigned int UINT16;
 #endif /* HAVE_UNSIGNED_SHORT */
 
-typedef long INT32;
+// L0 - Fix this
+//typedef long INT32;
+typedef signed int INT32;
+// End
 
 /* INT16 must hold at least the values -32768..32767. */
